Add project build for hatchling
Poetry is not working to build the project with direct dependencies, and
torch cannot be specified in a non-direct way as of today using this
dependency manager. Therefore, the build packages are also included in
pyproject.toml to be compatible with hatchling. The automated workflow
is corrected to use this build mechanism.