Add AutoLlamaIndex with basic configuration options
AutoLlamaIndex is a wrapper around LlamaIndex to easily create a vector
index and, given a query, call an LLM over the retrieved results.
LlamaIndex has many possible options regarding vector stores and LLMs,
and in this first implementation only a simple vector store is included
and only Hugging Face and OpenAI LLMs are accessible.