Add query with non-default models to AutoLlamaIndex
This functionality allows querying the database using an LLM different
to the one specified during the construction of the AutoLlamaIndex
object. It can be useful for API development.

Additionally, a new hybrid retriever is implemented based on the
combination of the results of two separate retrievers. As a first
approach, BM25 and vector retrievers are combined.